split welcome-tour.ts: extract onboarding state into onboarding-state.ts
